Rename all route files to prevent name collisions
