+++ kifi-backend/modules/abook/conf/abookService.routes
@@ -41,7 +41,7 @@ POST     /internal/abook/:userId/hideEmailFromUser              @com.keepit.aboo
 GET      /internal/abook/database/getEmailAccountsChanged       @com.keepit.abook.controllers.ABookDataPipeController.getEmailAccountsChanged(seqNum: SequenceNumber[com.keepit.abook.model.EmailAccount], fetchSize: Int)
 GET      /internal/abook/database/getContactsChanged            @com.keepit.abook.controllers.ABookDataPipeController.getContactsChanged(seqNum: SequenceNumber[com.keepit.abook.model.EContact], fetchSize: Int)
 
-->  / common.Routes
+->  / commonService.Routes
