+++ kifi-backend/modules/abook/conf/application-abook.conf
@@ -2,4 +2,4 @@
 include classpath("application-dev.conf")
 
 application.global="com.keepit.dev.ABookDevGlobal"
-application.router=abook.Routes
+application.router=abookService.Routes
