+++ kifi-backend/modules/common/conf/application.conf
@@ -1,2 +1,2 @@
-application.router=common.Routes
+application.router=commonService.Routes
 include "../../common/conf/application-dev.conf"
