+++ kifi-backend/modules/common/conf/prod/abook.conf
@@ -4,7 +4,7 @@ application.global="com.keepit.abook.ABookGlobal"
 
 include "prod.conf"
 
-application.router=abook.Routes
+application.router=abookService.Routes
 
 application.name="ABOOK"
