+++ kifi-backend/modules/common/conf/prod/c_shoebox.conf
@@ -7,7 +7,7 @@ include "prod.conf"
 # mark this as "canary" node
 service.canary=true
 
-application.router=shoebox.Routes
+application.router=shoeboxService.Routes
 
 application.name="C_SHOEBOX"
