+++ kifi-backend/modules/common/conf/prod/cortex.conf
@@ -4,7 +4,7 @@ include "prod.conf"
 
 
 application.global="com.keepit.cortex.CortexGlobal"
-application.router=cortex.Routes
+application.router=cortexService.Routes
 application.name="CORTEX"
 
 statsd {
