+++ kifi-backend/modules/common/conf/prod/curator.conf
@@ -4,7 +4,7 @@ include "prod.conf"
 
 
 application.global="com.keepit.curator.CuratorGlobal"
-application.router=curator.Routes
+application.router=curatorService.Routes
 application.name="CURATOR"
 
 statsd {
