+++ kifi-backend/modules/common/conf/prod/eliza.conf
@@ -4,7 +4,7 @@ application.global="com.keepit.eliza.ElizaGlobal"
 
 include "prod.conf"
 
-application.router=eliza.Routes
+application.router=elizaService.Routes
 
 application.name="ELIZA"
