+++ kifi-backend/modules/common/conf/prod/heimdal.conf
@@ -4,7 +4,7 @@ application.global="com.keepit.heimdal.HeimdalGlobal"
 
 include "prod.conf"
 
-application.router=heimdal.Routes
+application.router=heimdalService.Routes
 
 application.name="HEIMDAL"
