+++ kifi-backend/modules/common/conf/prod/scraper.conf
@@ -4,7 +4,7 @@ application.global="com.keepit.scraper.ScraperGlobal"
 
 include "prod.conf"
 
-application.router=scraper.Routes
+application.router=scraperService.Routes
 
 application.name="SCRAPER"
