+++ kifi-backend/modules/common/conf/prod/shoebox.conf
@@ -4,7 +4,7 @@ application.global="com.keepit.shoebox.ShoeboxGlobal"
 
 include "prod.conf"
 
-application.router=shoebox.Routes
+application.router=shoeboxService.Routes
 
 application.name="SHOEBOX"
