+++ kifi-backend/modules/cortex/conf/application-cortex.conf
@@ -2,4 +2,4 @@
 include classpath("application-dev.conf")
 
 application.global="com.keepit.dev.CortexDevGlobal"
-application.router=cortex.Routes
+application.router=cortexService.Routes
