+++ kifi-backend/modules/cortex/conf/cortexService.routes
@@ -1,5 +1,5 @@
 
-->  / common.Routes
+->  / commonService.Routes
 
 
 ##########################################
