+++ kifi-backend/modules/curator/conf/application-curator.conf
@@ -2,4 +2,4 @@
 include classpath("application-dev.conf")
 
 application.global="com.keepit.dev.CuratorDevGlobal"
-application.router=curator.Routes
+application.router=curatorService.Routes
