+++ kifi-backend/modules/curator/conf/curatorService.routes
@@ -1,5 +1,5 @@
 
-->  / common.Routes
+->  / commonService.Routes
 
 ##########################################
 # INTERNAL ROUTES/Curator
