+++ kifi-backend/modules/eliza/conf/application-eliza.conf
@@ -1,4 +1,4 @@
 include classpath("application-dev.conf")
 
 application.global="com.keepit.dev.ElizaDevGlobal"
-application.router=eliza.Routes
+application.router=elizaService.Routes
