+++ kifi-backend/modules/graph/conf/graphService.routes
@@ -9,4 +9,4 @@ GET     /internal/graph/getUserAndScorePairs @com.keepit.graph.controllers.inter
 GET     /internal/graph/getSociallyRelatedEntities  @com.keepit.graph.controllers.internal.GraphController.getSociallyRelatedEntities(userId: Id[User])
 POST    /internal/graph/explainFeed                      @com.keepit.graph.controllers.internal.GraphController.explainFeed
 
-->  / common.Routes
+->  / commonService.Routes
