+++ kifi-backend/modules/heimdal/conf/application-heimdal.conf
@@ -2,4 +2,4 @@
 include classpath("application-dev.conf")
 
 application.global="com.keepit.dev.HeimdalDevGlobal"
-application.router=heimdal.Routes
+application.router=heimdalService.Routes
