+++ kifi-backend/modules/scraper/conf/application-scraper.conf
@@ -2,4 +2,4 @@
 include classpath("application-dev.conf")
 
 application.global="com.keepit.dev.ScraperDevGlobal"
-application.router=scraper.Routes
+application.router=scraperService.Routes
