+++ kifi-backend/modules/shoebox/conf/application-shoebox.conf
@@ -2,4 +2,4 @@
 include classpath("application-dev.conf")
 
 application.global="com.keepit.dev.ShoeboxDevGlobal"
-application.router=shoebox.Routes
+application.router=shoeboxService.Routes
